refactor(security): extract secure_storage into ts-secure-storage crate (#91)
Вынос зашифрованного хранилища API-ключей (SecureStorage / ApiKeyType, AES-GCM поверх
tauri-plugin-store) из монолита src-tauri/security/secure_storage.rs в новый крейт
ts-secure-storage. Второй шаг к разблокировке ai_api_proxy (использует SecureStorage+ApiKeyType).

Файл полностью self-contained (ноль crate::-зависимостей), поэтому переезжает без правок логики:
- crates/ts-secure-storage: новый крейт (tauri/tauri-plugin-store/aes-gcm/anyhow/serde +
  serde_json/log/chrono/dirs; dev: tempfile/tokio). Добавлен в workspace members.
- git mv security/secure_storage.rs -> crates/ts-secure-storage/src/lib.rs.
- Монолит: security/secure_storage.rs стал ре-экспорт-шимом `pub use ts_secure_storage::*;` —
  security/mod.rs::`pub use secure_storage::*` и потребители (vision_commands, state,
  ai_api_proxy) резолвятся без правок. + path-dep в src-tauri/Cargo.toml.

Проверки: cargo test -p ts-secure-storage -> 7 passed; cargo check --lib (src-tauri) ->
0 ошибок, 3 baseline-варнинга.
